fixed bezier curve crash fixed image destructor bug fixed warnings

+++ vm/native/trivial/color.hpp
@@ -19,10 +19,10 @@ struct Color {
 	constexpr Color(float r, float g, float b, float a = 1.0f) : r(r), g(g), b(b), a(a) {}
 
 	uint32_t get_r8g8b8a8_unorm() const {
-		uint8_t r_u8 = r * 255;
-		uint8_t g_u8 = g * 255;
-		uint8_t b_u8 = b * 255;
-		uint8_t a_u8 = a * 255;
+		uint8_t r_u8 = static_cast<uint8_t>(r * 255.0);
+		uint8_t g_u8 = static_cast<uint8_t>(g * 255.0);
+		uint8_t b_u8 = static_cast<uint8_t>(b * 255.0);
+		uint8_t a_u8 = static_cast<uint8_t>(a * 255.0);
 
 		return (r_u8 << 24) | (g_u8 << 16) | (b_u8 << 8) | (a_u8 << 0);
 	}
